Corrected ext2hier.c so that resistance output from extresist takes the right argument type (float, not int).  Otherwise all resistances from extresist come out zero when doing "ext2spice extresist on" and "ext2spice hierarchy on".  Also changed the format of the resistance in the SPICE output to type float, since values are in standard units of ohms, and rounding to the nearest ohm seems excessively coarse-grained.

+++ resis/ResPrint.c
@@ -85,10 +85,10 @@ ResPrintExtRes(outextfile,resistors,nodename)
 	  }
 	  if (ResOptionsFlags & ResOpt_DoExtFile)
 	  {
-     	       fprintf(outextfile, "resist \"%s\" \"%s\" %d\n",
+     	       fprintf(outextfile, "resist \"%s\" \"%s\" %g\n",
 		    resistors->rr_connection1->rn_name,
 		    resistors->rr_connection2->rn_name,
-		    (int) (resistors->rr_value/ExtCurStyle->exts_resistScale));
+		    resistors->rr_value / (float)ExtCurStyle->exts_resistScale);
 	  }
      }
 }
